requirements.txt
requests>=2.31.0
mutagen>=1.47.0
Pillow>=10.0.0
pyperclip>=1.8.2
colorama>=0.4.6

